Move EFIFB out of HDMI.
