+++ platform/msm_shared/include/mdss_hdmi.h
@@ -35,5 +35,4 @@ uint32_t hdmi_pll_config(uint32_t tmds_clk_rate);
 int hdmi_vco_enable(void);
 int hdmi_vco_disable(void);
 void mdss_hdmi_display_init(uint32_t rev, void *base);
-void mdss_uefi_display_init(uint32_t rev);
 #endif
