+++ platform/msm_shared/rules.mk
@@ -268,6 +268,7 @@ DEFINES += DISPLAY_TYPE_MDSS=1
             $(LOCAL_DIR)/mipi_dsi_phy.o \
             $(LOCAL_DIR)/mipi_dsi_autopll.o \
             $(LOCAL_DIR)/mdss_hdmi.o \
+			$(LOCAL_DIR)/mdss_efifb.o \
             $(LOCAL_DIR)/hdmi_pll_28nm.o \
             $(LOCAL_DIR)/spmi.o \
             $(LOCAL_DIR)/bam.o \
@@ -485,6 +486,7 @@ DEFINES += DISPLAY_TYPE_MDSS=1
 			$(LOCAL_DIR)/mipi_dsi_autopll.o \
 			$(LOCAL_DIR)/mipi_dsi_autopll_20nm.o \
 			$(LOCAL_DIR)/mdss_hdmi.o \
+			$(LOCAL_DIR)/mdss_efifb.o \
 			$(LOCAL_DIR)/hdmi_pll_20nm.o \
 			$(LOCAL_DIR)/dload_util.o
 endif
